# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Resume Analyzer",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_api_auth_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/analysis/latest/{resume_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Latest Analysis",
        "description": "Return the most recent analysis result for a given resume.",
        "operationId": "get_latest_analysis_api_analysis_latest__resume_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resume Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/history/{resume_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Analysis History",
        "description": "Return all analysis results for a given resume, newest first.",
        "operationId": "get_analysis_history_api_analysis_history__resume_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resume Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/{analysis_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Analysis",
        "description": "Return a single analysis result by ID.",
        "operationId": "get_analysis_api_analysis__analysis_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/builder/save": {
      "post": {
        "tags": [
          "builder"
        ],
        "summary": "Save Built Resume",
        "description": "Saves the JSON built resume.",
        "operationId": "save_built_resume_api_builder_save_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeBuilderSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/builder/load": {
      "get": {
        "tags": [
          "builder"
        ],
        "summary": "Load Built Resume",
        "description": "Loads the user's saved JSON built resume.",
        "operationId": "load_built_resume_api_builder_load_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "AnalysisResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "resume_id": {
            "type": "string",
            "title": "Resume Id"
          },
          "target_role_id": {
            "type": "string",
            "title": "Target Role Id"
          },
          "readiness_score": {
            "type": "number",
            "title": "Readiness Score"
          },
          "skill_score": {
            "type": "number",
            "title": "Skill Score"
          },
          "project_score": {
            "type": "number",
            "title": "Project Score"
          },
          "professional_presence_score": {
            "type": "number",
            "title": "Professional Presence Score"
          },
          "strengths": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strengths"
          },
          "weaknesses": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weaknesses"
          },
          "recommendation_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Summary"
          },
          "analyzed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Analyzed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "resume_id",
          "target_role_id",
          "readiness_score",
          "skill_score",
          "project_score",
          "professional_presence_score",
          "analyzed_at"
        ],
        "title": "AnalysisResponse",
        "description": "Full analysis result returned to the client."
      },
...
      "Body_login_api_auth_login_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "password"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_api_auth_login_post"
      },
...
      "BuilderCertification": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "issuer": {
            "type": "string",
            "title": "Issuer"
          },
          "date": {
            "type": "string",
            "title": "Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "issuer",
          "date"
        ],
        "title": "BuilderCertification"
      },
      "BuilderEducation": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "institution": {
            "type": "string",
            "title": "Institution"
          },
          "degree": {
            "type": "string",
            "title": "Degree"
          },
          "year": {
            "type": "string",
            "title": "Year"
          }
        },
        "type": "object",
        "required": [
          "id",
          "institution",
          "degree",
          "year"
        ],
        "title": "BuilderEducation"
      },
      "BuilderExperience": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "startDate": {
            "type": "string",
            "title": "Startdate"
          },
          "endDate": {
            "type": "string",
            "title": "Enddate"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company",
          "role",
          "startDate",
          "endDate",
          "description"
        ],
        "title": "BuilderExperience"
      },
      "BuilderPersonalInfo": {
        "properties": {
          "fullName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fullname",
            "default": ""
          },
          "jobTitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jobtitle",
            "default": ""
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "default": ""
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone",
            "default": ""
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "default": ""
          },
          "linkedin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin",
            "default": ""
          },
          "portfolio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio",
            "default": ""
          },
          "photo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo"
          }
        },
        "type": "object",
        "title": "BuilderPersonalInfo"
      },
      "BuilderProject": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "link": {
            "type": "string",
            "title": "Link"
          },
          "tools": {
            "type": "string",
            "title": "Tools"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "link",
          "tools",
          "description"
        ],
        "title": "BuilderProject"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResumeBuilderSaveRequest": {
        "properties": {
          "personalInfo": {
            "$ref": "#/components/schemas/BuilderPersonalInfo"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "experience": {
            "items": {
              "$ref": "#/components/schemas/BuilderExperience"
            },
            "type": "array",
            "title": "Experience"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/BuilderEducation"
            },
            "type": "array",
            "title": "Education"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/BuilderProject"
            },
            "type": "array",
            "title": "Projects"
          },
          "certifications": {
            "items": {
              "$ref": "#/components/schemas/BuilderCertification"
            },
            "type": "array",
            "title": "Certifications"
          }
        },
        "type": "object",
        "required": [
          "personalInfo",
          "summary",
          "skills",
          "experience",
          "education",
          "projects",
          "certifications"
        ],
        "title": "ResumeBuilderSaveRequest"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "role",
          "user_id"
        ],
        "title": "Token"
      },
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "full_name",
          "role",
          "password"
        ],
        "title": "UserCreate"
      },
      "UserResponse": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "email",
          "full_name",
          "role",
          "id",
          "is_active",
          "created_at"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}